Catch image rescale issues with small dtypes
For arrays that are being sent to the image saving functions, if they have small dtypes like bool_ (true/false), int8 (-128 to 127), and uint8 (0 to 255) then trying to rescale them may not be safe and you'll end up with wrong values.  E.G. trying to rescale an int8 array holding 0s and 1s (as in a black & white mask) to values in the 0 to 255 range ends up with values of -1 and 127 because int8 can't hold a value of 255.

Catching these and turning them into arrays with a different dtype will allow for safe rescaling. float64 may be overkill but we can optimize that later.  Technically for the above code int8 is safe because presently we are rescaling to a range of 0 to 255 which is what in8 holds, but as the code comment mentions we also want to work with 16-bit image data in the near future which will require a range of 0 to 65535, so we will be needing to catch and convert int8 arrays as well.